backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0
alembic>=1.13.0
pydantic-settings>=2.1.0
livekit-agents>=0.8.0
livekit-api>=0.6.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
livekit-plugins-google>=0.8.0
livekit-plugins-silero>=0.6.0
livekit-plugins-cartesia>=0.4.0
cartesia>=1.0.0
pyjwt[crypto]>=2.8.0
httpx>=0.26.0
slowapi>=0.1.9
jinja2>=3.1.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
opik>=1.0.0
